Refactor Create person
