+++ app.rb
@@ -3,13 +3,15 @@ require './teacher'
 require './book'
 require './rental'
 require './modules/menu'
-require './modules/list_book'
-require './modules/list_people'
+require './modules/list_all_book_1'
+require './modules/list_all_people_2'
+require './modules/create_a_person_3'
 
 class App
   include Menu
   include ListBooks
   include ListPeople
+  include CreatePerson
 
   OPERATIONS = {
     1 => :list_books,
@@ -41,44 +43,6 @@ class App
         puts 'That is not a valid input'
       end
     end
-  end
-
-  def create_person
-    print 'Do you want to create a student (1) or a teacher (2)? [Input the number]: '
-    option = gets.chomp.to_i
-    case option
-    when 1
-      create_student
-    when 2
-      create_teacher
-    else
-      puts 'That is not a valid input'
-    end
-  end
-
-  def create_student
-    print 'Age: '
-    age = gets.chomp.to_i
-    print 'Name: '
-    name = gets.chomp
-    print 'Has parent permission? [Y/N]: '
-    parent_permission = gets.chomp
-    parent_permission = parent_permission.downcase == 'y'
-    student = Student.new(age, name, parent_permission)
-    puts 'Person created successfully'
-    @people << student
-  end
-
-  def create_teacher
-    print 'Age: '
-    age = gets.chomp.to_i
-    print 'Name: '
-    name = gets.chomp
-    print 'Specialization: '
-    specialization = gets.chomp
-    teacher = Teacher.new(age, specialization, name)
-    puts 'Person created successfully'
-    @people << teacher
   end
 
   def create_book
